% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Example_IDEAM.R
\docType{data}
\name{Example_IDEAM}
\alias{Example_IDEAM}
\title{A dataset with fictitious values of no real IDEAM's Stations.  The text file
 keeps IDEAM's text format.}
\format{text file}
\source{
It is a self-made file. Original data could be freely required through
IDEAM's web page. \url{http://www.ideam.gov.co/solicitud-de-informacion}
}
\usage{
data(Example_IDEAM)
}
\description{
A dataset with fictitious values of no real IDEAM's Stations.  The text file
 keeps IDEAM's text format.
}
\examples{
data(Example_IDEAM)
}
\keyword{datasets}
