% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/IdeamWide.R
\name{IdeamWide}
\alias{IdeamWide}
\title{Create a matrix with IDEAM's data}
\usage{
IdeamWide(file, write = FALSE, outfile = "MatrizIdeamAncha")
}
\arguments{
\item{file}{IDEAM file path or file name if it is on the working directory}

\item{write}{If True a CSV file is returned to the working directory, otherwise
only a data frame is shown}

\item{outfile}{Outfile name that will be saved on the working directory}
}
\value{
a data frame or a CSV file
}
\description{
Create a data frame that it is available to export as a CSV file
}
\examples{
# Retreive example dataset
Example_IDEAM <- system.file("extdata", "Example_IDEAM", package = "ideamdb")
# Create a temporal file
example.ideam.wide <- tempfile()
write.csv(IdeamWide(Example_IDEAM), file = example.ideam.wide)
read.csv(example.ideam.wide)

}
