% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TimeSeries.R
\name{TimeSeries}
\alias{TimeSeries}
\title{Graphical exploratory charts}
\usage{
TimeSeries(file, station = "all", variable = "all")
}
\arguments{
\item{file}{IDEAM file path or file name if it is on the working directory}

\item{station}{a single station name or type "all" or "todas" to plot everything}

\item{variable}{a single variable name or type "all" or "todas" to plot everything
only a data frame is shown}
}
\value{
time series plot and boxplot for every variable data and station
}
\description{
Create a time series chart and a boxplot of every data sample
}
\examples{
Example_IDEAM <- system.file("extdata", "Example_IDEAM", package = "ideamdb")
TimeSeries(Example_IDEAM, station = "LUCERO")
# Print time series chart and boxplots for all variables
# collected in an specific station

}
